app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.11.0

Flask==2.2.2
Flask-Cors==3.0.10
requests==2.28.1
google-api-python-client==2.71.0
bs4==0.0.1
numpy==1.24.1
python-dotenv==0.21.0
scipy==1.10.0
transformers==4.25.1